backend ml module

+++ Backend/Application/AIML/ProductData.cs
@@ -1,17 +1,17 @@
+using Microsoft.ML.Data;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+
 
 namespace Application.AIML
 {
     public class ProductData
     {
+        
         public string Type { get; set; }
         public string Name { get; set; }
-        public string Description { get; set; }
-        public float Price { get; set; }
-        public int Review { get; set; }
+        public string Description { get; set; }    
+        public float Price { get; set; }  
+        public float Review { get; set; }
     }
 }
